Bind mount Exosphere files in docker image, rather than copying
Bind mounting files in the Docker image means that files on the local
disk and in the Docker image cannot get out of sync. This makes for a
better developer experience, as whatever source is present on the host
machine will be used when running the Docker image.

As this requires an inordinately long Docker command, we add a script
to start the docker image, and update documentation to use that
script.

Fixes #547