Show error message when no flavors are accessible in the current project
Now that flavors are modeled using RemoteDataPlusPlus, we can
distinguish between a project which has no flavors and one which
simply has not had flavors loaded yet. This allows us to render the
"Create Instance" page based primarily on the state of the project --
we no longer need to track implicitly in the currently selected flavor
whether or not project flavors have been loaded. Note that this
removes the need for the `GotFlavorList` message.

The "selected flavor" on the "Create Instance" page now represents
whether or not the user has selected a flavor explicitly. On initial
page load, we select the smallest flavor by default, which is the
existing behavior.

Fixes #950