Fix Python syntax warning
Starting in Python 3.12, regular expressions containing "special sequences" must use raw string notation (i.e. begin with an `r`) to avoid a syntax warning. [0] [1] We started to see this warning from system_load_json.py in Ubuntu 24.04, which ships with Python 3.12. This change adopts the raw string notation to fix the warning.

I expect this change to be backward-compatible with our oldest supported operating systems, going back to Python 3.6 and older. [2]

Big thanks to community member Kyle Mohr for reporting this (and providing the fix!) in Jetstream2 public chat. [3]

[0] https://docs.python.org/3/whatsnew/3.12.html

[1] https://docs.python.org/3.12/library/re.html

[2] https://docs.python.org/3.6/library/re.html

[3] https://matrix.to/#/!elSIKLFMgTdJRyjiwr:matrix.org/$0Rwhc6Og7pQE08Fms0I2XxpwdaimpMhCab5z6BtXVFE